Comment display & delete
Created a post action to clear the post errors object
and put it into the errorReducer so as to be used from
everywhere.

Created an post action to delete a comment that takes
into consideration if the comment's author is the one
who is logged in to be able to delete the comment.